requirements.txt
python-chess
zstandard
polars
tqdm
plotnine